More refactor, output "unrounded" times.

+++ fast/interop.c
@@ -205,8 +205,8 @@ void timer_end(i64 start) {
   // print time elapsed time
   i64 end = timer_start();
   printf("End of time is: %.2f \n", end * 1e-3);
-  float elapsed_time = (end - start);
-  printf("Elapsed time is: %.2f secs\n", ((float) elapsed_time) * 1e-3);
+  i64 elapsed_time = (end - start);
+  printf("Elapsed time is: %.17f secs\n", elapsed_time * 1e-3);
 }
